Fix: require tax ID when purchase overrides threshold
git-svn-id: http://plugins.svn.wordpress.org/woocommerce-quaderno/trunk@1703586 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/class-wc-qd-invoice-manager.php
@@ -77,7 +77,7 @@ class WC_QD_Invoice_Manager {
 		}
 
 		// Let's create the receipt or the invoice
-		if ( $order->get_billing_country() == WC_Countries::get_base_country() && $order->get_total() < intval( WC_QD_Integration::$receipts_threshold ) && empty( $vat_number ) && empty( $tax_id ) ) {
+		if ( $order->get_billing_country() == WC_Countries::get_base_country() && $order->get_total() < floatval( WC_QD_Integration::$receipts_threshold ) && empty( $vat_number ) && empty( $tax_id ) ) {
 		  $invoice = new QuadernoReceipt($invoice_params);
 		} else {
 		  $invoice = new QuadernoInvoice($invoice_params);
